requirements.txt
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
pydantic[email]>=2.5.0
pydantic-settings>=2.1.0
motor>=3.3.2
pymongo>=4.6.1
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
bcrypt==4.0.1
python-multipart>=0.0.6
jinja2>=3.1.3
pytest>=7.4.0
pytest-asyncio>=0.21.0
httpx>=0.25.0
